Trying to get this thing back in the store 1.0
The only problem I see is that the search suggestions are not showing
up. In a prior version, users could start typing and be given a list of
5 or so locations.

Also, I think it would be better if the two buttons were always visible
and not only on hover. Touch users cannot ever see those buttons.